move contract calls into ContractProvider and signer load into AccountProvider
